Move AreaItem to its own class
